+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## {{ UNRELEASED_VERSION }} - [{{ UNRELEASED_DATE }}]({{ UNRELEASED_LINK }})
 
+## v23089235556.31.1-build.63e25bb58041f15d45ffb32417026e25447f9594 - [March 14, 2026](git://github.com/tanaabased/prepare-release-action.git)
+
 ## v23060923011.30.1-build.a5519099e15563e7183cfa11802d9621223e1034 - [March 13, 2026](git://github.com/tanaabased/prepare-release-action.git)
 
 ## v23060722186.29.1-build.b6d0d8dac6795e4e2e636f66323abd9a3c862e4b - [March 13, 2026](git://github.com/tanaabased/prepare-release-action.git)